test: allow dependency injection via client module
